requirements.txt
bibtexparser==1.4.3
biopython==1.85
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
idna==3.10
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
